requirements.txt
ezc3d
numpy
flask
gunicorn
python-dotenv